Code cleanup and portrait support
